tests/e2e/poc-dapr-patterns/service-a/requirements.txt
dapr>=1.14.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
fastapi>=0.109.0
uvicorn>=0.27.0
tenacity>=8.2.0
